requirements.txt
# Core ML
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
peft>=0.4.0

# Medical imaging
nibabel>=5.0.0
SimpleITK>=2.2.0
antspyx>=0.3.8
hd-bet>=2.0

# Data and visualization
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.2.0
scipy>=1.10.0
Pillow>=9.5.0
tqdm>=4.65.0

# Interpretability
grad-cam>=1.4.8